packages/clew-ingestion/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "clewso-ingestion"
version = "0.1.6"
description = "Ingestion worker for Clew"
requires-python = ">=3.11"
dependencies = [
    "qdrant-client>=1.7.0",
    "neo4j>=5.14.0",
    "gitpython>=3.1.40",
    "tree-sitter>=0.22,<1.0",
    "tree-sitter-language-pack>=0.6",
    "openai>=1.6.0",
    "pydantic>=2.5.0",
    "tenacity>=8.2.3",
    "httpx>=0.27.0",
    "clewso-core>=0.1.0",
]

[project.urls]
Homepage = "https://github.com/clewnet/clewso"
Repository = "https://github.com/clewnet/clewso"

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "respx>=0.20.2",
]

[tool.uv.sources]
clewso-core = { workspace = true }

[tool.hatch.build.targets.wheel]
packages = ["src/clewso_ingestion"]
